Move add.rb to the ruby/tests directory
